Add elevation profile downsampling

+++ src/main/java/com/trailmatch/dto/ElevationProfilePointDto.java
@@ -1,7 +1,3 @@
 package com.trailmatch.dto;
 
-/**
- * Point du profil altimétrique exposant l'altitude à une distance cumulée donnée.
- */
-public record ElevationProfilePointDto(double distanceKm, int elevationM) {
-}
+public record ElevationProfilePointDto(double distanceKm, double elevationM) {}
